R9 y R10 (1-2): contrato de mremap, y checks derivados del perfil
R9. `sys_mremap` documenta y aplica el contrato: flags distintos de 0 siguen
sin soportarse, la dirección va alineada, una longitud cero es EINVAL (no un
no-op) y las longitudes se normalizan al alza a página como en mmap/munmap.
`grow_anon` reserva el intervalo bajo el candado del libro de regiones y
materializa las páginas con el candado suelto —el fault-in reentra en el
gestor de memoria—; si falla a medias, el rollback devuelve la longitud
anterior y retira solo las páginas que mapeó la operación. Una página ya
presente en el hueco aborta sin adoptarla.

Los casos se comprueban con llamadas reales en QEMU (`init test`): cero,
desalineación, no-op inválido, crecer desde el interior, flags, solape (con la
región vecina intacta en contenido y permisos) y un hijo `mremap-oom` que pide
más páginas de las que hay —tamaño calculado con `meminfo`, así que falla con
cualquier RAM— y verifica que tras el rollback la región sigue midiendo lo
mismo, con su contenido y sus permisos.

R10.1. `cargo xtask check` ejecuta *todos* los tests de xtask; el filtro
anterior dejaba fuera `pci_stable::` y `elf_mmap_rules::`. Cada paso guarda su
salida en `target/check-artefactos/` y la vuelca al fallar. Los bancos C de
GSP y ath11k se compilan y ejecutan también con ASan+UBSan, como ya hacía el
de iwlwifi; eso destapó un desplazamiento con UB en el fixture de ath11k.

R10.2. Los hostchecks obligatorios se derivan del perfil compilado: si el
perfil trae ath11k, su banco no se puede omitir porque falte el script, y se
comprueba que el firmware que ese port necesita está en el rootfs salvo que el
perfil lo excluya. El perfil por defecto conserva la exigencia previa
(iwlwifi + nouveau) y el perfil mínimo sigue sin exigir nada.

+++ tools/ath11k-hostcheck/main.c
@@ -149,7 +149,8 @@ static void sim_procesa_descarga(void)
     uint32_t base = SIM_BHIE_OFF;
     if (sim.fw_falla) {
         sim.reg[(base + BHIE_TXVECSTATUS_OFFS) / 4] =
-            (BHIE_TXVECSTATUS_STATUS_ERROR << BHIE_TXVECSTATUS_STATUS_SHIFT) | sim.fw_seq;
+            ((uint32_t)BHIE_TXVECSTATUS_STATUS_ERROR << BHIE_TXVECSTATUS_STATUS_SHIFT) |
+            sim.fw_seq;
         return;
     }
     const struct bhi_vec_entry *vec = (const struct bhi_vec_entry *)(uintptr_t)sim.vec_addr;
@@ -162,7 +163,8 @@ static void sim_procesa_descarga(void)
     sim.fw_bytes_vistos = total;
     sim.fw_descargado = 1;
     sim.reg[(base + BHIE_TXVECSTATUS_OFFS) / 4] =
-        (BHIE_TXVECSTATUS_STATUS_XFER_COMPL << BHIE_TXVECSTATUS_STATUS_SHIFT) | sim.fw_seq;
+        ((uint32_t)BHIE_TXVECSTATUS_STATUS_XFER_COMPL << BHIE_TXVECSTATUS_STATUS_SHIFT) |
+        sim.fw_seq;
     /* Tras cargar amss.bin el dispositivo cambia de entorno. */
     sim.ee = MHI_EE_AMSS;
     sim.reg[(SIM_BHI_OFF + BHI_EXECENV) / 4] = MHI_EE_AMSS;
